requirements.txt
numpy
pandas
python-dotenv
pymongo
certifi
scikit-learn
pymongo[srv]==3.6
##-e .
